pyproject.toml
[project]
name = "dms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appdirs==1.4.4",
    "asgiref==3.11.1",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "cryptography==46.0.5",
    "diff-match-patch==20241021",
    "django==5.1.6",
    "django-allauth==65.14.3",
    "django-autoslug==1.9.9",
    "django-environ==0.12.1",
    "django-import-export==4.4.0",
    "django-safedelete==1.4.1",
    "django-simple-history==3.11.0",
    "django-widget-tweaks==1.5.1",
    "fs==2.4.16",
    "idna==3.11",
    "inflection==0.5.1",
    "num2words>=0.5.14",
    "oauthlib==3.3.1",
    "openpyxl>=3.1.5",
    "packaging==26.0",
    "pillow==12.1.1",
    "pycparser==3.0",
    "pyjwt==2.11.0",
    "python-slugify==8.0.4",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "setuptools==79.0.1",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "tablib==3.9.0",
    "text-unidecode==1.3",
    "tzdata==2025.3",
    "unicode-slugify==0.1.5",
    "unidecode==1.4.0",
    "urllib3==2.6.3",
    "uv==0.7.13",
    "whitenoise==6.11.0",
]
